ALL KINDS OF CHANGES: every task has its own file, authors names are corrected, cleaner imports etc.
